+++ packages/terra-divider/src/Divider.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import classNames from 'classnames/bind';
 import 'terra-base/lib/baseStyles';
-import styles from './Divider.scss';
+import styles from './Divider.module.scss';
 
 const cx = classNames.bind(styles);