+++ packages/terra-form-fieldset/src/Fieldset.jsx
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 
 import 'terra-base/lib/baseStyles';
-import styles from './Fieldset.scss';
+import styles from './Fieldset.module.scss';
 
 const cx = classNames.bind(styles);